pyproject.toml
[tool.poetry]
name = "power_manager"
version = "0.1.0"
description = ""
authors = ["ktomi96 <kokai.tamas.96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
midea-beautiful-air = "^0.9.15"
autopep8 = "^2.0.1"
python-dotenv = "^0.21.0"
flask = "^2.2.2"
livereload = "^2.6.3"
jinja2 = "^3.1.2"
flask-wtf = "^1.0.1"
apscheduler = "^3.9.1.post1"
pandas = "^1.5.2"
py-solaredge = "^0.0.4.9"
plotly = "^5.11.0"
memoization = "^0.4.0"
sqlalchemy = "^1.4.46"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
